refactor(ui): extract keyboard shortcuts and inset hooks into keyboard module
