+++ src/ui/keyboard/mod.rs
@@ -1,0 +1,7 @@
+mod inset;
+#[cfg(target_os = "macos")]
+mod macos;
+
+pub use inset::use_keyboard_inset;
+#[cfg(target_os = "macos")]
+pub use macos::use_macos_shortcuts;
